feat(about): new footer section on about us page
