--- src/components/Standalone/styles.css
@@ -1,78 +1,0 @@
-h2 {
-  @mixin font-serif;
-
-  font-size: 1.7rem;
-  line-height: 1.4;
-
-  @media (--sm-up) {
-    font-size: 2.5rem;
-  }
-}
-
-.footer {
-  padding: 4.5rem 0 2rem;
-  text-align: center;
-  background: var(--color-grey-lighter);
-
-  & .footer-section {
-    margin-bottom: var(--spacing-xxx-loose);
-  }
-
-  & h2 {
-    margin-bottom: var(--spacing-loose);
-    font-size: var(--font-size-lg);
-  }
-
-  & p {
-    @mixin font-serif;
-
-    & + p {
-      margin-top: var(--spacing-loose);
-    }
-  }
-
-  & .socials {
-    & :global(> *) {
-      display: inline-block;
-      margin: 0 calc(var(--spacing-loose) / 2);
-    }
-  }
-
-  & form {
-    display: flex;
-
-    & input {
-      @mixin border-bottom-grey;
-
-      display: inline-block;
-      height: 2.5rem;
-      padding: var(--spacing-x-tight);
-      margin-right: var(--spacing-base);
-      background: var(--color-white);
-    }
-  }
-
-  & .copyright {
-    @mixin font-serif;
-
-    font-size: var(--font-size-sm);
-    color: var(--color-grey-dark);
-  }
-
-  @media (--sm-up) {
-    padding: 7.5rem 0;
-    text-align: left;
-
-    & .socials {
-      & :global(> *) {
-        margin-right: var(--spacing-loose);
-        margin-bottom: var(--spacing-x-tight);
-        margin-left: 0;
-      }
-    }
-
-    & .copyright {
-      color: var(---grey-dark);
-    }
-  }
-}
