+++ src/views/About/index.tsx
@@ -1,11 +1,11 @@
 import Link from 'next/link'
 
 import { Head, IconLogo } from '~/components'
-import Footer from '~/components/Standalone/Footer'
 
 import { PATHS, TEXT } from '~/common/enums'
 
 import Features from './Features'
+import Footer from './Footer'
 import Goal from './Goal'
 import Reports from './Reports'
 import Slogan from './Slogan'
